add user token

+++ controllers/blogs.js
@@ -1,9 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const { extractUser } = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
+    .populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
@@ -16,13 +18,16 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const user = await User.findById(request.body.userId)
+
+blogsRouter.post('/', extractUser, async (request, response) => {
+  const body = request.body
+  const user = request.user
+
   const blog = new Blog({
-    title: request.body.title,
-    author: request.body.author,
-    url: request.body.url,
-    likes: request.body.likes,
+    title: body.title,
+    author: user.name,
+    url: body.url,
+    likes: body.likes,
     user: user.id
   })
 
@@ -38,7 +43,15 @@ blogsRouter.put('/:id', async (request, response) => {
   response.status(200).json(updatedEntry)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', extractUser, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(400).json({ error: 'no blog' })
+  }
+  if (blog.user.toString() !== request.user.id) {
+    return response.status(401).json({ error: 'cannot delete unowned blog' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
